Replace requests-html with httpx and beautifulsoup4
The library `requests-html` is more or less a wrapper around the
`requests` and `beautifulsoup4` libraries but has the disadvantage of
being less flexible and doesn't have type hints. It also hasn't been
very well maintained recently. Therefore, I'm removing the dependency on
that library.

We lose one feature: being able to easily load pages with the wrapped
`pyppeteer` library. I'm not even sure we were using it before but, in
any case, it should be possible to add that as an optional dependency in
the future, for example to handle pages with links rendered by
JavaScript.

This also removes the use of `requests` in favor of `httpx`. The former
doesn't have native support for asynchronous I/O or HTTP2, and seems
less flexible overall. We'll see if this turns out to be a good library.
It should be easy to revert that change should it turn out to be a bad
one.

I configured `beautifulsoup4` with `http.parser` instead of `lxml`, as
was done by `requests-html` if I understand correctly. It's one less
dependency. Again, we'll see if it works well, otherwise it will be
trivial to reintroduce `lxml`.

Lastly, this adds a integration test with an HTTP redirect. I added it
because `httpx` doesn't follow redirects by default and I want to make
sure we keep the same behavior as with `requests`, at least for now.
However, the new test case highlights a potential issue with redirects
(which already existed before): the crawler remembers the source URL,
not the destination URL, which may be confusing. I'll investigate that
in a future commit/pull request.