Improve internal representation of pages
This transposes the internal representation. Instead of storing a map of
links and pages where they were found, we store a map of pages and links
they contain.

For now, this doesn't change any visible behavior of the program,
because we still use the same visible representation for the user, but
that may change in the future.

This new representation is already easier to reason about, because it's
closer to HTTP and HTML pages, and will hopefully enable new use cases
in the near future, like reporting redirects clearly to the user.