Transpose the returned page/link structure
Instead of returning a map of {link → pages where the link was found},
we return a map of {page → links found on the page}.

This is a continuation of an earlier commit
(2a84da1906f8e0fdee065b5417dc3708729a850d) which only changed the
internal representation. The present commit changes the external
representation, that is, what is returned to the user as text or JSON.

Hopefully, this representation is more intuitive and will enable more
use cases like the representation of redirects.