Store and show HTTP redirects
Redirects were previously handled implicitly and ignored, which means
that the source URL would often be considered to be the location of a
page. For example, if `example.net/foo` redirected to `example.net/bar`,
it would consider `/foo` to be a page with links, which is kind of
wrong.

Instead, redirects are followed just like links on a page and
intermediate results are stored in the store. The only exception is the
starting point, which is still handled a bit differently.

Thanks to this change, we avoid visiting redirection targets if we've
visited them before. This wasn't the case before.

The output format (text or JSON) was also changed. It now shows the
redirection path for each link on a page.