Rename `pages` to `url_infos` in `UrlStore`
Similarly as in the previous commit, we fix some names to better reflect
the data they refer to. In this case, it's the internal data structure
which no longer just contains information about pages, but about URLs
and request results in general (not just pages but also redirects and
connection errors).