Handle unknown URL schemes
The parser previously explicitly excluded the `mailto` scheme to avoid
crashing when encountering such a scheme in an `href` attribute.
Unfortunately, this approach doesn't work given all the potential
schemes to be observed (see
https://en.wikipedia.org/wiki/List_of_URI_schemes for more information).

As a result, the parser now only accepts listed schemes (it uses an
allow list instead of a deny list). That should make it much more
robust.